+++ src/vue/modules/requests/create-asset-requests/module.js
@@ -3,10 +3,10 @@ import { ModuleDescriptor } from '@/modules-arch/module-descriptor'
 export class CreateAssetRequestsModule extends ModuleDescriptor {
   constructor (opts = {}) {
     super({
-      importComponentFn: _ => import('@/vue/modules/requests/create-asset-requests'),
+      importComponentFn: _ => import('@modules/requests/create-asset-requests'),
       importStoreFn: async _ => {
         const { createAssetRequestsModule: res } =
-          await import('@/vue/modules/requests/create-asset-requests/store')
+          await import('@modules/requests/create-asset-requests/store')
         return res
       },
       ...opts,